feat(rapid):refine predicate impl
1: refine the impl of predicate.
2: to support vectorized version of predicate evaluation.

Issues:#622
